requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
streamlit
python-dotenv
ipwhois
requests
